pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-control-plane"
version = "0.1.0"
description = "MCP control plane server"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "MCP Control Plane" }]
dependencies = [
  "mcp>=1.0.0",
  "pydantic>=2.6",
  "prometheus-client>=0.20",
  "opentelemetry-api>=1.24",
  "opentelemetry-sdk>=1.24",
  "opentelemetry-exporter-otlp>=1.24",
  "opentelemetry-instrumentation>=0.45b0",
  "uvicorn>=0.29",
  "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
  "mypy>=1.10",
  "coverage[toml]>=7.5",
  "types-setuptools",
]

[project.scripts]
mcp-control-plane = "mcp_cp.server:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC", "RUF"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
mypy_path = ["src"]
ignore_missing_imports = true
disallow_subclassing_any = false

[[tool.mypy.overrides]]
module = [
  "httpx",
  "mcp.*",
  "opentelemetry",
  "opentelemetry.*",
  "prometheus_client.*",
  "prometheus_client",
  "pydantic",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["mcp_cp"]

[tool.coverage.report]
fail_under = 85
show_missing = true
